Fix Console Errors

+++ src/item.h
@@ -856,9 +856,10 @@ class Item : virtual public Thing
 		int32_t getSlotPosition() const {
 			return items[id].slotPosition;
 		}
+
 		int8_t getHitChance() const {
 			if (hasAttribute(ITEM_ATTRIBUTE_HITCHANCE)) {
-				return getIntAttr(ITEM_ATTRIBUTE_HITCHANCE);
+				return static_cast<uint8_t>(getIntAttr(ITEM_ATTRIBUTE_HITCHANCE));
 			}
 			return items[id].hitChance;
 		}
